+++ mdsvex.config.js
@@ -1,9 +1,9 @@
-import { defineMDSveXConfig as defineConfig } from 'mdsvex';
+import { defineMDSveXConfig as defineConfig } from "mdsvex";
 
 const config = defineConfig({
-	extensions: ['.svelte.md', '.md', '.svx'],
+	extensions: [".svelte.md", ".md", ".svx"],
 	smartypants: {
-		dashes: 'oldschool',
+		dashes: "oldschool",
 	},
 	remarkPlugins: [],
 	rehypePlugins: [],