+++ src/lib/index.ts
@@ -1,5 +1,5 @@
-import type { ComponentProps } from 'svelte';
-import Ripple from './Ripple.svelte';
+import type { ComponentProps } from "svelte";
+import Ripple from "./Ripple.svelte";
 
 export { Ripple };
 export type RippleProps = ComponentProps<Ripple>;