+++ src/routes/+layout.svelte
@@ -1,9 +1,9 @@
 <script>
-	import 'prismjs';
-	import 'prismjs/components/prism-bash';
-	import 'prism-svelte';
-	import '../app.css';
-	import '../prism.css';
+	import "prismjs";
+	import "prismjs/components/prism-bash";
+	import "prism-svelte";
+	import "../app.css";
+	import "../prism.css";
 </script>
 
 <main class="prose mx-auto max-w-3xl p-8">