+++ svelte.config.js
@@ -1,13 +1,13 @@
-import { preprocessMeltUI } from '@melt-ui/pp';
-import adapter from '@sveltejs/adapter-static';
-import { vitePreprocess } from '@sveltejs/kit/vite';
-import { mdsvex } from 'mdsvex';
-import sequence from 'svelte-sequential-preprocessor';
-import mdsvexConfig from './mdsvex.config.js';
+import { preprocessMeltUI } from "@melt-ui/pp";
+import adapter from "@sveltejs/adapter-static";
+import { vitePreprocess } from "@sveltejs/kit/vite";
+import { mdsvex } from "mdsvex";
+import sequence from "svelte-sequential-preprocessor";
+import mdsvexConfig from "./mdsvex.config.js";
 
 /** @type {import('@sveltejs/kit').Config}*/
 const config = {
-	extensions: ['.svelte', ...mdsvexConfig.extensions],
+	extensions: [".svelte", ...mdsvexConfig.extensions],
 	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
 	// for more information about preprocessors
 	preprocess: sequence([vitePreprocess({}), mdsvex(mdsvexConfig), preprocessMeltUI()]),
@@ -17,7 +17,7 @@ const config = {
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter(),
 		alias: {
-			$docs: './src/docs',
+			$docs: "./src/docs",
 		},
 	},
 };